refactor: useMyProfile로직 분리
