+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useLogout';
 export * from './useLike';
 export * from './useIntersect';
 export * from './useHorizontalDragScroll';
+export * from './useMyProfile';
